Use vim-dirvish instead of netrw
Netrw is significantly slow in NeoVim on macOS, and that's my primary
OS. So, after posting it here:
https://www.reddit.com/r/neovim/comments/btsmpb/netrw_is_considerably_slower_in_neovim_than_vim/
somebody suggested vim-dirvish and I like the way it works and its
speed. It's also much more aligned with Vim-like interface. So, I'm
disabling netrw and using vim-dirvish from now on. Let's see how it
goes.